Ch05 bug fixes
Fixed a bug with wrong tooltip information for scrolls

Fixed a bug when the player based inventory items are still in the store inventory when exiting. They are now removed.

Fixed a bug where the items were not being saved to the correct open slots in the player inventory

Fixed crash when loading non-existent profile from the load screen

If we run out of room when buying items and still left items to be sold, then we will stack remaining items in the last slot.